+++ ViewModels/SettingsViewModel.cs
@@ -157,6 +157,7 @@ internal sealed partial class SettingsViewModel : ObservableObject
             {
                 OcrEngineKind.Paddle => "Paddle",
                 OcrEngineKind.PaddleVllm => "PaddleVllm",
+                OcrEngineKind.Ndl => "Ndl",
                 _ => "WinRt"
             };
             PaddleDetectionModelName = settings.PaddleTextDetectionModelName;
@@ -266,6 +267,7 @@ internal sealed partial class SettingsViewModel : ObservableObject
         {
             "Paddle" => OcrEngineKind.Paddle,
             "PaddleVllm" => OcrEngineKind.PaddleVllm,
+            "Ndl" => OcrEngineKind.Ndl,
             _ => OcrEngineKind.WinRt
         };
         settings.PaddleTextDetectionModelName = PaddleDetectionModelName;